refactor: change package name to match application name
